extract board view model interface
