+++ project/src/androidTest/java/de/earthlingz/oerszebra/DroidZebraTest.java
@@ -3,7 +3,7 @@ package de.earthlingz.oerszebra;
 import android.content.Intent;
 import android.test.ActivityInstrumentationTestCase2;
 import com.shurik.droidzebra.ZebraEngine;
-import de.earthlingz.oerszebra.BoardView.BoardViewModel;
+import de.earthlingz.oerszebra.BoardView.GameStateBoardModel;
 
 /**
  * This is a simple framework for a test of an Application.  See
@@ -86,7 +86,7 @@ public class DroidZebraTest extends ActivityInstrumentationTestCase2<DroidZebra>
     }
 
     private int countSquares(byte color) {
-        BoardViewModel state = this.getActivity().getState();
+        GameStateBoardModel state = this.getActivity().getState();
         int result = 0;
         for (int y = 0, boardLength = state.getBoardHeight(); y < boardLength; y++) {
             for (int x = 0, rowLength = state.getBoardRowWidth(); x < rowLength; x++) {
