+++ project/src/main/java/de/earthlingz/oerszebra/ZebraServices.java
@@ -4,15 +4,16 @@ import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 
 import de.earthlingz.oerszebra.BoardView.BoardViewModel;
+import de.earthlingz.oerszebra.BoardView.GameStateBoardModel;
 import de.earthlingz.oerszebra.parser.GameParser;
 import de.earthlingz.oerszebra.parser.ReversiWarsParser;
 import de.earthlingz.oerszebra.parser.ThorParser;
 
 class ZebraServices {
-    private static final Supplier<BoardViewModel> boardState = Suppliers.memoize(BoardViewModel::new);
+    private static final Supplier<GameStateBoardModel> boardState = Suppliers.memoize(GameStateBoardModel::new);
     private static final Supplier<GameParser> gameParser = Suppliers.memoize(() -> new GameParser(new ReversiWarsParser(), new ThorParser()));
 
-    public static BoardViewModel getBoardState() {
+    public static GameStateBoardModel getBoardState() {
         return boardState.get();
     }
